♻️ Refactor : LoginResponse DTO 폴더로 이동 및 Schema title 추가
- LoginResponse 클래스를 domain/auth/dto 폴더로 이동
- LoginResponse, UserInfoDto 클래스의 @Schema title 수정 및 명확한 설명 추가
