+++ src/main/java/com/umc7th/a1grade/domain/auth/service/OAuth2TokenService.java
@@ -1,6 +1,6 @@
 package com.umc7th.a1grade.domain.auth.service;
 
-import com.umc7th.a1grade.domain.auth.controller.response.LoginResponse;
+import com.umc7th.a1grade.domain.auth.dto.LoginResponse;
 import com.umc7th.a1grade.domain.auth.dto.GoogleTokenResponse;
 import com.umc7th.a1grade.domain.auth.dto.OAuthAttributes;
