+++ src/main/java/com/umc7th/a1grade/domain/user/dto/UserInfoDto.java
@@ -13,6 +13,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
+@Schema(title = "UserInfoDto - 닉네임 입력을 위한 데이터 전송 객체")
 public class UserInfoDto {
   @NotBlank(message = "닉네임은 필수 입력 항목입니다. 최대 5자 까지 입력 가능")
   @Size(max = 5)
